Referrer url changes - Included first referrer page, changed view count and unique view count

+++ api/src/main/java/com/visumbu/api/adwords/report/xml/bean/AccountDeviceReportRow.java
@@ -38,7 +38,7 @@ public class AccountDeviceReportRow {
 
     public String getAvgCPC() {
         try {
-            return Integer.toString(Integer.parseInt(avgCPC) / 1000000);
+            return Double.toString(Double.parseDouble(avgCPC) / 1000000);
         } catch (Exception e) {
             return "0";
         }
@@ -60,7 +60,7 @@ public class AccountDeviceReportRow {
 
     public String getCost() {
         try {
-            return Integer.toString(Integer.parseInt(cost) / 1000000);
+            return Double.toString(Double.parseDouble(cost) / 1000000);
         } catch (Exception e) {
             return "0";
         }
@@ -199,7 +199,7 @@ public class AccountDeviceReportRow {
 
     public String getCostConv() {
         try {
-            return Integer.toString(Integer.parseInt(costConv) / 1000000);
+            return Double.toString(Double.parseDouble(costConv) / 1000000);
         } catch (Exception e) {
             return "0";
         }
